refactor: move openapi-validator to yarn workspace
This is a long overdue cleanup of our package management. The common
code was in an openapi-validator folder duplicated in the jest and chai
packages. Version 0.9.0 of the openapi-validator plugin had been
published to npm but then no longer updated. The jest and chai packages
listed openapi-validator 0.9.0 in their package.jsons, so were
downloading it but actually used their local copies instead, which devs
had to update manually. Moreover, openapi-validator had its own
package.json that was unused; we relied on the top level yarn.lock
instead. It was a big fragile mess.

This deduplicates the openapi-validator folder and migrates it to a yarn
workspace package, so the package.json is now correctly used. The chai
and jest plugins will use the local code in dev, and the published npm
package in prod, meaning openapi-validator now needs to be deployed
before the chai and jest plugins are deployed.